+++ src/main/java/io/enderdev/endermodpacktweaks/config/CfgFeatures.java
@@ -36,6 +36,40 @@ public class CfgFeatures {
                 "[Wither] ore:proofWither"
         })
         public boolean enable = false;
+
+        @Config.RequiresMcRestart
+        @Config.Name("[02] Default Wither")
+        @Config.Comment("These blocks will be unable to be broken by the wither.")
+        public String[] defaultWither = new String[]{
+                "minecraft:bedrock",
+                "minecraft:end_portal",
+                "minecraft:end_portal_frame",
+                "minecraft:command_block",
+                "minecraft:repeating_command_block",
+                "minecraft:chain_command_block",
+                "minecraft:barrier",
+                "minecraft:structure_block",
+                "minecraft:structure_void",
+                "minecraft:piston_extension",
+                "minecraft:end_gateway"
+        };
+
+        @Config.RequiresMcRestart
+        @Config.Name("[03] Default Ender Dragon")
+        @Config.Comment("These blocks will be unable to be broken by the ender dragon.")
+        public String[] defaultEnderDragon = new String[]{
+                "minecraft:bedrock",
+                "minecraft:end_portal",
+                "minecraft:end_portal_frame",
+                "minecraft:command_block",
+                "minecraft:repeating_command_block",
+                "minecraft:chain_command_block",
+                "minecraft:barrier",
+                "minecraft:obsidian",
+                "minecraft:end_stone",
+                "minecraft:iron_bars",
+                "minecraft:end_gateway"
+        };
     }
 
     @Config.Name("forced_resourcepack")